% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort_initial.R
\name{create_cohort_initial}
\alias{create_cohort_initial}
\title{Helper Function: Create new cohort initially}
\usage{
create_cohort_initial(
  trial_struc,
  cohorts_start,
  n_int,
  n_fin,
  rr_comb_vec,
  rr_mono_vec,
  rr_back_vec,
  rr_plac_vec
)
}
\description{
Helper Function: Create new cohort initially
}
