% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohort_new.R
\name{create_cohort_new}
\alias{create_cohort_new}
\title{Helper Function: Create new cohort and update allocation ratio}
\usage{
create_cohort_new(
  res_list,
  plac,
  n_int,
  n_fin,
  sharing_type,
  rr_comb_vec,
  rr_mono_vec,
  rr_back_vec,
  rr_plac_vec
)
}
\description{
Helper Function: Create new cohort and update allocation ratio
}
