% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sr_reached.R
\name{is_sr_reached}
\alias{is_sr_reached}
\title{Helper Function: Check whether stopping rule reached}
\usage{
is_sr_reached(res_list, sr_drugs_pos, sr_pats, expected)
}
\description{
Helper Function: Check whether stopping rule reached
}
