% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prob_bin.R
\name{post_prob_bin}
\alias{post_prob_bin}
\title{Helper function posterior probability of one Beta being by margin delta greater than other Beta}
\usage{
post_prob_bin(
  n_exp,
  n_contr,
  resp_exp,
  resp_contr,
  delta,
  a0_exp,
  b0_exp,
  a0_contr,
  b0_contr
)
}
\description{
Helper function posterior probability of one Beta being by margin delta greater than other Beta
}
